Move Request pojo into api package
