+++ sormas-e2e-tests/src/test/java/org/sormas/e2etests/helpers/RestAssuredClient.java
@@ -29,7 +29,7 @@ import io.restassured.specification.RequestSpecification;
 import javax.inject.Inject;
 import javax.inject.Named;
 import lombok.SneakyThrows;
-import org.sormas.e2etests.pojo.web.Request;
+import org.sormas.e2etests.pojo.api.Request;
 import org.sormas.e2etests.state.ApiState;
 
 public class RestAssuredClient {
